+++ Runtime/Math/Extensions.cs
@@ -9,5 +9,15 @@ namespace DotsTween.Math
         {
             return new float4(color.r, color.g, color.b, color.a);
         }
+
+        /// <summary>
+        /// Truncates alpha channel.
+        /// </summary>
+        /// <param name="color"></param>
+        /// <returns></returns>
+        public static float3 ToFloat3(this Color color)
+        {
+            return new float3(color.r, color.g, color.b);
+        }
     }
 }